+++ app/controllers/variables_controller.rb
@@ -20,6 +20,14 @@ class VariablesController < ApplicationController
   end
 
   def show
+    graphs_folder = File.join(@dataset.root_folder, 'dd', 'graphs')
+    valid_folders = if File.exist?(graphs_folder)
+      (Dir.entries(graphs_folder) - ['.','..']).select{|f| !File.file?(f)}
+    else
+      []
+    end
+    @version = (valid_folders.include?(params[:v].to_s.strip) ? valid_folders.select{|f| f == params[:v].to_s.strip}.first : @variable.version.to_s)
+
     render layout: 'nonavigation'
   end